==> What Movie Should I Watch Next?/01 - Pulling the Top Rated Movies Dataset Using TMDB API.ipynb <==
## 1. Importing Python Libraries

We shall start by importing the essential Python libraries

### IMPORTING LIBRARIES
import pandas as pd
import numpy as np
import requests
import collections

## 2. Connecting to TMDB Website Using the API Key

We now use the API key we had received when we signed up for TMDB to pull data from the website. For this, we go to https://developers.themoviedb.org and on the left column, we can see a section of urls that can help us get the information we need. We go to _'Get Top Rated'_ under _MOVIES_ and select the url to get the list of top rated movies and attach our API key as directed. 

Here, we will only try to get the first page of results so we mention page number as 1 in the url. We then create a response object by passing the url through requests. 

### USING API KEY TO GET DATA
api_key = "API Key"
url = "https://api.themoviedb.org/3/movie/top_rated?api_key=" +  api_key + "&language=en-US&page=1"
response = requests.get(url)
response

A response 200 means our request was a success. 

## 3. Pulling the Top Rated Movies Data from the Response Object I

We can now extract information from this response object. We can see from https://developers.themoviedb.org that the json file in the response object contains an array _results_ which has the information we need. Let us make a pandas dataframe, _tmdb_df_, from this array.

### CREATING A PANDAS DATAFRAME 
tmdb_df = pd.DataFrame(response.json()['results'])
tmdb_df

We can see that the movie data contains a lot of information about the top rated movies: their certification, genres, unique ids, original title, title in English, plot summary, release date and so on. We actually wont be needing all of these columns but we shall deal with that later. 

## 4. Creating a Dictionary of Column Names

Before moving forward, let us store the names of all the columns in a dictionary. This will make our task easier down the line.

### CREATING A DICTIONARY OF COLUMN NAMES
feature_names = collections.defaultdict(str)
for name in tmdb_df.columns:
    feature_names[name]
feature_names

## 5. Pulling the Top Rated Movies Data from the Response Object II

Here, we shall create a for-loop that will increment page numbers from 2 to any number that we want(here we take the number as 469). Using the same process as above, we shall extract the array contained in the json file of the response object obtained from each of these pages, convert them into a pandas dataframe and vertically stack them on the dataframe _tmbd_df_.

### PULLING DATA FOR ALL THE TOP RATED MOVIES
for page_no in range(2, 469):
    url = "https://api.themoviedb.org/3/movie/top_rated?api_key=" +  api_key + "&language=en-US&page= %d" % page_no
    print('Getting page: %d' % page_no)
    response = requests.get(url)
    tmdb_results = response.json()['results']
    movie_data = pd.DataFrame(tmdb_results)
    tmdb_df = np.vstack((tmdb_df, movie_data))

We then convert the final array into a pandas dataframe and name the columns using the dictionary that we had previously created.

### CREATING A PANDAS DATAFRAME
tmdb_df = pd.DataFrame(tmdb_df, columns = feature_names)
tmdb_df

## 6. Saving the Movies Dataframe

Now, let us save this dataframe in csv format which can be used later to make models.

### SAVING THE DATAFRAME
tmdb_df.to_csv('tmdb_movies_data.csv', index = False)
