==> What Movie Should I Watch Next?/02 - Pulling the Director and Cast Data for the Top Rated Movies.ipynb <==
## 1. Importing Python Libraries

We shall start by importing the essential Python libraries

### IMPORTING LIBRARIES
import numpy as np
import pandas as pd
import requests

## 2. Importing and Cleaning the Top Rated Movies Dataframe

Now, we shall create a variable for the API key and also import the movies dataframe that we had pulled in our last session.

api_key = "API Key"

movies = pd.read_csv('tmdb_movies_data.csv')
movies.head(10)

Next, we will be dropping some of the columns that contain information which are not helpful to us.

### DROPPING UNNECESSARY COLUMNS
movies.drop(['backdrop_path', 'original_title', 'overview', 'poster_path'], axis = 1, inplace = True)
movies.head(10)

Now, in order to perform content based filtering, the amount of information present in the dataframe _movies_ is not enough. To establish similarity between two movies, we need features that can be used to relate them. Here, we will try to add more information by adding the name of the director of each of these movies and the five main actors starring in each of them. Hence, if we choose one movie, the next recommended movie will have either a similar cast or the same director.


Now, to store data properly and to make searches efficient, TMDB gives a unique id to each movie. We can see the column _'id'_ in _movies_ which contains such ids. Let us extract these ids separately for easy use.

movie_ids = movies['id']

## 3. Pulling Director Data Using API Key

We now use our API key to pull crew data from the website. For this, we go to https://developers.themoviedb.org and on the left column, we can see a section of urls that can help us get the information we need. We select the url for _'Get Credits'_ under _MOVIES_ and attach the movie id and API key as directed. We then extract information from the response object that we get. We can see from https://developers.themoviedb.org that the json file in the response object contains an array _crew_ which has the information we need. Let us make a pandas dataframe, _tmdb_crew_, from this array, then extract the name of the director of that movie. 

We shall loop the process over all the movie ids using a for-loop and then make a list containing the movie ids in one column and its respective director in the other.

### EXTRACTING DIRECTORS FOR EVERY TOP RATED MOVIE
tmdb_direct = []
for movie_id in movie_ids:
    url = "https://api.themoviedb.org/3/movie/%d" % movie_id + "/credits?api_key=" + api_key + "&language=en-US"
    response = requests.get(url)
    tmdb_crew = pd.DataFrame(response.json()['crew']) #extracting array from the json file
    director = tmdb_crew[tmdb_crew['job'] == 'Director']['name'].values.tolist()[0] #extracting director name
    tmdb_direct.append([movie_id, director])

Let us make a pandas dataframe from this list.

### CREATING A PANDAS DATAFRAME CONTAINING MOVIE IDS AND DIRECTOR NAMES
tmdb_direct = pd.DataFrame(tmdb_direct, columns = ['id', 'director'])
tmdb_direct

## 4. Saving the Director Dataframe

Then we store this dataframe in csv format for future use.

tmdb_direct.to_csv('tmdb_directors.csv', index = False)

## 5. Pulling Cast Data Using API Key


Next, we try to do the same thing to get information about the actors appearing in the movies. Since there are numerous actors in a movie, we shall only consider the top 5 leads. We shall use our API key to pull cast data from the website. For this, we go to https://developers.themoviedb.org and on the left column, we can see a section of urls that can help us get the information we need. We select the url for _'Get Credits'_ under _MOVIES_ and attach the movie id and API key as directed. We then extract the information from the response object that we get. We can see from https://developers.themoviedb.org that the json file in the response object contains an array _cast_ which has the information we need. Let us make a pandas dataframe, _tmdb_cast_, from this array.

Here, we face two problems:
1. Some movies do not contain any information about its cast i.e. they lack the column containing information about the cast. To work around this problem, we shall apply a filter which will only allow movies having exactly 12(maximum number of columns) columns to pass through.
2. Some movies have less than 5 actors. For this, we shall make the following if-else statement: If the number of actors is more than 5, then we only consider the top 5 leads; if it is less than 5, then we shall extend the list by adding 5 'None' values and then consider the top 5 leads.  

We shall loop the process over all the movie ids using a for-loop and then make a list containing the movie ids in one column and the top 5 leads in the other 5 columns.

### EXTRACTING THE CAST FOR EACH TOP RATED MOVIE
tmdb_act = []
for movie_id in movie_ids:
    url = "https://api.themoviedb.org/3/movie/%d" % movie_id + "/credits?api_key=" + api_key + "&language=en-US"
    response = requests.get(url)
    tmdb_cast = pd.DataFrame(response.json()['cast'])
    if(len(tmdb_cast.columns) == 12): #filtering cast data with exactly 12 columns
        actors = tmdb_cast[tmdb_cast['known_for_department'] == 'Acting']['name'].tolist() #extracting the cast data
        if(len(actors) >= 5):
            actors = actors[0:5] #extracting the five main lead actors
        else:
            actors.extend(['None', 'None', 'None', 'None', 'None']) #extending the cast list
            actors = actors[0:5] #extracting the five main lead actors
    tmdb_act.append([movie_id, actors[0], actors[1], actors[2], actors[3], actors[4]])

We shall convert this list into a pandas dataframe.

### CREATING A PANDAS DATAFRAME CONTAINING MOVIE ID AND THE FIVE MAIN ACTORS
tmdb_act = pd.DataFrame(tmdb_act, columns = ['id', 'actor_1', 'actor_2', 'actor_3', 'actor_4', 'actor_5'])
tmdb_act

## 6. Saving the Cast Dataframe

And we store this dataframe in csv format for future use.

tmdb_act.to_csv('tmdb_actors.csv', index = False)
