==> What Movie Should I Watch Next?/03 - Pulling the Data for Movie Genres.ipynb <==
## 1. Importing Python Libraries

We shall start by importing the essential Python libraries

### IMPORTING LIBRARIES
import numpy as np
import pandas as pd
import requests
import re

## 2. Importing the Top Rated Movies Dataframe

Now, we shall create a variable for the API key and also import the movies dataframe that we had pulled in the first session.

api_key = "API Key"
movies = pd.read_csv('tmdb_movies_data.csv')
movies.head(10)

## 3. Manipulating the Genre IDs

Now, let us consider the genre of the first movie.

genre = movies['genre_ids'][0]
genre

We see that although it has a bunch of genre ids, it is given as a string rather than a list of integers.

len(genre)

Here, the length should have been 3 but it's returning a value of 15 further proving that it exists as a string value. Now, let us try to convert this string into a list of integers. For this, we first get rid of the two inverted commas at the beginning and at the end of the string. We then split the remaining string using commas and convert the remaining values as integers to get a list.

genre = genre[1:len(genre)-1] #removing the inverted commas
genre = list(genre.split(", ")) #splitting by commas to get integers
genre = [int(item) for item in genre] #converting to get a lost of integers
genre

We can check to see if we indeed were able to convert the string to a list.

len(genre)

We've successfully obtained a list of genre ids. 

## 4. Extracting the Genre IDs for All the Top Rated Movies

Next, let us try to do this for all the movies using a loop. So for each movie in the list, we clean and split the genre ids.

Here, we face two problems:
1. Some movies have a blank space as genre id. To solve this, we shall replace the blank space with 0.
2. After converting the string to a list, we see that some movies have more than 2 genre ids. For this, we shall only take the top 2 genre ids. If a movie has less than 2 genre ids, then we shall extend the list by adding two 0 values and then consider the top 2 genre ids.

We shall then make a list containing the two genre ids in each column.

### GETTING GENRE IDS FOR ALL THE TOP RATED MOVIES
tmdb_genre_keys = []
for movie in range(0, len(movies)):
    genre_key = movies['genre_ids'][movie] 
    genre_key = genre_key[1:len(genre_key)-1] #removing the inverted commas
    genre_key = list(genre_key.split(', ')) #splitting by commas to get integers
    for index, item in enumerate(genre_key): 
        if(item == ''):
            genre_key[index] = 0 #replacing each blank genre id with 0 
    genre_key = [int(float(item)) for item in genre_key] #converting to get a list of integers
    gen_length = len(genre_key)
    if(gen_length < 2): 
        genre_key.extend([0, 0]) #extending the genre id list
        genre_key = genre_key[0:2] #extracting the top two genre ids
    elif(gen_length >= 2):
        genre_key = genre_key[0:2] #extracting the top two genre ids
    tmdb_genre_keys.append([genre_key[0], genre_key[1]]) 

We then convert this list into a pandas dataframe _tmdb_genre_keys_.

### CREATING A PANDAS DATAFRAME FOR GENRE IDS
tmdb_genre_keys = pd.DataFrame(tmdb_genre_keys, columns = ['genre_1', 'genre_2'])
tmdb_genre_keys

## 5. Pulling the Genre ID Tags Using the API Key

Next, we shall use our API key to pull genre id tags from the TMDB website. For this, we go to https://developers.themoviedb.org and on the left column, we can see a section of urls that can help us get the information we need. We go to the _GENRES_ section then select the url for _'Get Movie List'_ and attach the API key as directed. We then extract information from the response object we get. We can see from the https://developers.themoviedb.org that the json file in the response object contains an array which has the information we need. Let us make a pandas dataframe _tmdb_genre_list_ from this array.

### EXTRACTING TAGS FOR MOVIE GENRE IDS
url = "https://api.themoviedb.org/3/genre/movie/list?api_key=" +  api_key + "&language=en-US"
response = requests.get(url)
tmdb_genre_list = pd.DataFrame(response.json())
tmdb_genre_list

## 6. Cleaning the Genre ID Tags Data

Now, let us clean this data so that we can get a dataframe which has genre ids in one column and the corresponding genre tag in the other column. For this, we shall do the following:

1. Convert each row value as a string.
2. Remove all punctuations.
3. Remove the term _'id'_ by splitting the string using that term and taking the second value.
4. Remove the term _'name'_ from the remaining string by splitting it using that term which will leave us with genre id as first value and genre name as the second value.
5. Convert the genre id into an integer value and append the id and name to a list.

We shall do this for all the genres in _tmdb_genre_list_ by using a for-loop. Lastly, since we had added 0 when either a movie had missing genre id or had less than 2 genres, we shall also add another row with genre id as 0 and genre name as _'None'_.

### CLEANING THE TAGS DATA
tmdb_genre_tags = []
for genre_index in range(0, len(tmdb_genre_list)):
    genre = tmdb_genre_list['genres'][genre_index]
    genre = str(genre) #1
    genre = re.sub(r'[^\w\s]+', '', genre) #2
    genre = list(genre.split('id '))[1] #3
    genre = list(genre.split(' name ')) #4
    tmdb_genre_tags.append([int(float(genre[0])), genre[1]]) #5
tmdb_genre_tags.append([0, 'None']) #adding the genre tag for 0

Let us convert this list into a pandas dataframe _tmdb_genre_tags_.

### CREATING A PANDAS DATAFRAME FOR GENRE ID TAGS 
tmdb_genre_tags = pd.DataFrame(tmdb_genre_tags, columns = ['genre_id', 'genre_names'])
tmdb_genre_tags

## 7. Replacing the Genre IDs with their Respective Tags

First, we take the two columns of _tmdb_genre_ and store these values in two separate lists: _genre_keys_ and _genre_names_.

### STORING THE COLUMNS OF GENRE ID TAG DATAFRAME
genre_keys = list(tmdb_genre_tags['genre_id'])
genre_names = list(tmdb_genre_tags['genre_names'])

Then, we shall take the dataframe _tmdb_genre_keys_ and in each of its columns, we replace the genre id with the genre name using the lists _genre_keys_ and _genre_names_.

### REPLACING IDS WITH THEIR RESPECTIVE TAGS
tmdb_genre_keys['genre_1'] = tmdb_genre_keys['genre_1'].replace(genre_keys, genre_names)
tmdb_genre_keys['genre_2'] = tmdb_genre_keys['genre_2'].replace(genre_keys, genre_names)
tmdb_genre_keys

## 8. Saving the Genres Dataframe

Lastly, we store this dataframe in csv format.

tmdb_genre_keys.to_csv('tmdb_genres.csv', index = False)
