==> What Movie Should I Watch Next?/04 - Movie Recommendation with Content Based Filtering I.ipynb <==
## 1. Importing Python Libraries

We shall start by importing the essential Python libraries

### IMPORTING LIBRARIES
import numpy as np
import pandas as pd
import re
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics.pairwise import cosine_similarity

## 2. Importing the Datasets

Next, we shall import the datasets that we had pulled from TMDB.

### IMPORTING DATASETS
tmdb_movies = pd.read_csv('tmdb_movies_data.csv')
tmdb_genres = pd.read_csv('tmdb_genres.csv')
tmdb_directors = pd.read_csv('tmdb_directors.csv')
tmdb_actors = pd.read_csv('tmdb_actors.csv')

## 3. Cleaning and Merging the Datasets 

### a) Cleaning the Top Rated Movies Dataframe

First, let us take a look at _tmdb_movies_.

tmdb_movies

The dataframe contains 14 columns; some of them like _backdrop_path_, _'poster_path'_ and so on, contain no useful information. So, we shall drop such uninformative columns to get _tmdb_movies_updated_.

### DROPPING UNNECESSARY COLUMNS
tmdb_movies_updated = tmdb_movies.drop(['adult', 'backdrop_path', 'genre_ids', 'original_title', 'poster_path', 'video'], axis = 1)
tmdb_movies_updated

Next, let us take the title of each of these movies, convert them into string format and get rid of any punctuation. 

### CLEANING THE MOVIE TITLES
new_title = []
for name in range(0, len(tmdb_movies_updated)):
    title = tmdb_movies_updated['title'][name]
    title = str(title) #converting to string format
    title = re.sub(r'[^\w\s]+', '', title) #removing punctuations
    new_title.append(title)
tmdb_movies_updated['title'] = new_title
tmdb_movies_updated

### b) Merging with the Genres Dataframe

Lets take a look at the _tmdb_genres_ dataframe. 

tmdb_genres

We now combine _tmdb_movies_updated_ and _tmdb_genres_ by simply attaching the columns of _tmdb_genres_ to _tmdb_movies_updated_.

### COMBINING THE TOP RATED MOVIES AND MOVIE GENRE IDS DATASETS
tmdb_movies_updated['genre_1'] = tmdb_genres['genre_1']
tmdb_movies_updated['genre_2'] = tmdb_genres['genre_2']
tmdb_movies_updated

### c) Merging with the Director and Cast Dataframes

Next, we have the dataframe _tmdb_directors_.

tmdb_directors

And _tmdb_actors_.

tmdb_actors

Notice that the three dataframes _tmdb_movies_updated_, _tmdb_directors_ and _tmdb_actors_ all share the same column containing the unique ids for top rated movies. Thus, we can use inner join with respect to this unique id column to combine these three dataframes.

First, let us merge _tmdb_movies_updated_ and _tmdb_directors_.

### MERGING THE UPDATED MOVIES AND THE TMDB DIRECTOR DATASETS
tmdb_movies_updated = tmdb_movies_updated.merge(tmdb_directors, left_on = 'id', right_on = 'id')
tmdb_movies_updated

Then, we merge the resulting _tmdb_movies_updated_ with _tmdb_actors_ to get get _df_. 

### MERGING THE UPDATED MOVIES AND THE TMDB ACTORS DATASETS
df = tmdb_movies_updated.merge(tmdb_actors, left_on = 'id', right_on = 'id')
df

## 4. Creating a Combined Feature

In order to establish a relationship between any two movies, we need to get a measure that quantifies their relatedness. This is provided by calculating similarity scores which measures distances between two movies with dimensions representing their features. If the distance is small, the features are having a high degree of similarity. Whereas a large distance will be a low degree of similarity.

Let us first create a feature that is a combination of other features. Here, we shall take the features _'title'_, _'original_language'_, _'genre_1'_, _'genre_2'_, _'director'_, _'actor_1'_, _'actor_2'_, _'actor_3'_, _'actor_4'_ and _'actor_5'_ to create _'combined_features'_. So if we feed a movie to our model, the next recommended movie will possibly be of the same language, of similar genre and either have a similar cast or the same director.

### COMBINING FEATURES FOR SIMILARITY SCORES
def combine_features(data):
    return data['title'] + ' ' + data['original_language'] + ' ' + data['genre_1'] + ' ' + data['genre_2']  + ' '+ data['director'] + ' ' + data['actor_1'] + ' ' + data['actor_2'] + ' ' + data['actor_3'] + ' ' + data['actor_4'] + ' ' + data['actor_5']
df['combined_features'] = df.apply(combine_features, axis = 1)
df

## 5. Calculating Similarity Scores

Next, we shall create a Document Term Matrix(DTM) from this _'combined_features'_ column, where each row will represent a particular movie and the columns will represent all the words appearing in the _'combined_features'_ column of all the movies. Thus, each _(i,j)_-th entry of this DTM will be the frequency of appearance of the _j_-th word in the _'combined_features'_ column of the _i_-th movie. For this, we pass the _'combined_features'_ column through _CountVectorizer_.

We then feed this DTM into the _cosine_similarity_ function to get the similarity scores.

### CREATING A DTM AND CALCULATING SIMILARITY SCORES
vectorizer = CountVectorizer()
count = vectorizer.fit_transform(df['combined_features']) #creating dtm
similarity = cosine_similarity(count)

## 6. Creating Functions for Easy Flow of Data

Before going forward, let us take a moment to create two functions which will help us down the line:
1. _get_id_from_title_, which will take a movie title and return its unique id.
2. _get_title_from_id_, which will take a movie id and return its title.

### CREATION OF FUNCTIONS
def get_id_from_title(title):
    return df[df['title'] == title]['id'].values[0]

def get_title_from_id(movie_id):
        return tmdb_movies[tmdb_movies['id'] == movie_id]['title'].values[0]

## 7. Getting Movie Recommendations I

Now, we shall feed in a movie name so that we can get our recommendations. We choose _Spider-Man_ as the title we want recommendations for. We first remove any punctuation from the title and then get its unique id from the TMDB dataset.

### CLEANING THE USER DEFINED MOVIE TITLE
movie_user = 'Spider-Man'
movie_user = re.sub(r'[^\w\s]+', '', movie_user) #removing pupnctuations
movie_index = get_id_from_title(movie_user) #getting the unique id for the user defined movie
movie_index

Next, we try to obtain the row index of this movie in _df_ and then use this index to get the row of similarity scores between this particular movie and the remaining others. We transpose this row into a column and create a pandas dataframe, _df_similarity_. Next, we attach the unique movie ids from _df_ to _df_similarity_.

### GETTING SIMILARITY SCORES FOR THE USER DEFINED MOVIE
df_similarity = pd.DataFrame(similarity[df[df['id'] == movie_index].index].reshape(-1, 1), columns = ['cosine_similarity'])
df_similarity['id'] = df['id']
df_similarity

Now, we sort this dataframe with respect to the decreasing similarity scores.

### SORTING THE DATAFRAME IN DESCENDING ORDER OF SIMILARITY SCORES
df_similarity = df_similarity.sort_values(by = 'cosine_similarity', ascending = False, ignore_index = True)
df_similarity.head(10)

Lastly, we take the id of the top 20 movies with highest similarity scores and use them to get the titles of the recommended movies. These are the movies that are recommended by the model.

### PRINTING THE TOP 20 RECOMMENDED MOVIES
for movie in range(1, 21):
    movie_id = df_similarity['id'][movie]
    print(get_title_from_id(movie_id))

The recommendations are good but what we would like here is to improve the model so that other Spider-Man movies outside the Raimi trilogy are included in our recommendations.

## 8. Getting Movie Recommendations II

First, we make a function that automates the whole process described above: takes a movie title, calculates the similarity scores, and returns 20 recommendations.

def get_recommendations(movie_user):
    movie_user = re.sub(r'[^\w\s]+', '', movie_user) #removing punctuations
    movie_index = get_id_from_title(movie_user) #getting the unique id for the user defined movie
    df_similarity = pd.DataFrame(similarity[df[df['id'] == movie_index].index].reshape(-1, 1), columns = ['cosine_similarity'])
    df_similarity['id'] = df['id']
    df_similarity = df_similarity.sort_values(by = 'cosine_similarity', ascending = False, ignore_index = True)
    for movie in range(1, 21):
        movie_id = df_similarity['id'][movie]
        print(get_title_from_id(movie_id))

Let us now try to get recommendations for _The Dark Knight_. 

get_recommendations('The Dark Knight')

While the recommendations do contain the rest of the movies from the Nolan Trilogy, we would like to see more Batman movies there.
